Imported Upstream version 4.1.0

+++ src/fdt/ptx_twomasks.cc
@@ -80,7 +80,7 @@ void twomasks()
 
   ////////////////////////////////
   //  Log& logger = LogSingleton::getInstance();
-  volume<int> seeds,seeds2;
+  volume<float> seeds,seeds2;
   read_volume(seeds,opts.seedfile.value());
   read_volume(seeds2,opts.mask2.value());
